+++ docs/conf.py
@@ -44,6 +44,8 @@ nitpick_ignore = [
     ("py:class", "cwt.exceptions.CWTError"),
     ("py:class", "cwt.recipient.Recipient"),
     ("py:class", "CBORTag"),
+    ("py:class", "Claims"),
+    ("py:class", "cwt.claims.Claims"),
 ]
 
 # Add any paths that contain templates here, relative to this directory.